requirements.txt
fastapi
uvicorn
google-genai
pillow
requests
python-dotenv